Update log api
