--- src/apps/logs/api/serializers_singleton.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from apps.logs import models
-
-
-class ReviewLogSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.ReviewLog
-        fields = "__all__"
