+++ src/apps/logs/urls.py
@@ -1,6 +1,9 @@
 from django.urls import path
 from django.conf.urls import include
+from .api import api
 
 app_name = "logs"
 
-urlpatterns = [path("api/", include("apps.logs.api.urls", namespace="api"))]
+urlpatterns = [
+    path("api/", include(api.urls)),
+]
