+++ src/apps/surveys19/urls.py
@@ -1,5 +1,6 @@
 from django.urls import path
 from django.conf.urls import include
+from .api import api
 
 from .views import Surveys2019Index
 
@@ -7,5 +8,5 @@ app_name = "surveys19"
 
 urlpatterns = [
     path("", Surveys2019Index.as_view(), name="surveys19_index"),
-    path("api/", include("apps.surveys19.api.urls", namespace="api")),
+    path("api/", include(api.urls)),
 ]
