+++ src/templates/surveys19/index.html
@@ -8,11 +8,11 @@
 
 {% block external_css %}
 <link href="{% static 'vendor/pace/pace-theme-loading-bar.css' %}" rel="stylesheet">
-<link href="{% static 'css/surveys18/index.css' %}" rel="stylesheet">
+<link href="{% static 'css/index.css' %}" rel="stylesheet">
 {% endblock external_css %}
 
 {% block external_js %}
-<script src="{% static 'js/surveys19/index.js' %}"></script>
+<script src="{% static 'js/index.js' %}"></script>
 <script src="{% static 'js/surveys19/helper.js' %}"></script>
 {% endblock external_js %}
 
@@ -43,9 +43,9 @@
                                placeholder="請輸入農戶編號">
                         <div class="btn-group-vertical flex mt-1" role="group" aria-label="" style="width:100%;">
                             <button type="button" class="btn btn-success btn-lg js-get-survey"
-                                    data-readonly="false" data-url="{% url 'surveys19:api:survey-list' %}">查詢</button>
+                                    data-readonly="false" data-url="{% url 'surveys19:survey-list' %}">查詢</button>
                             <button type="button" class="btn btn-success btn-lg js-get-survey"
-                                    data-readonly="true" data-url="{% url 'surveys19:api:survey-list' %}">查詢原始資料</button>
+                                    data-readonly="true" data-url="{% url 'surveys19:survey-list' %}">查詢原始資料</button>
                         </div>
                     </div>
                 </div>
@@ -60,7 +60,7 @@
                         <a href="#" class="list-group-item js-tabs-control small" data-target="#panel3">戶內人口組成狀況</a>
                         <a href="#" class="list-group-item js-tabs-control small" data-target="#panel4">農業外僱人力</a>
                         <button type="button" class="btn btn-info btn-lg btn-block mt-1 js-set-survey"
-                                    data-url="{% url 'surveys19:api:survey-list' %}">更新調查表</button>
+                                    data-url="{% url 'surveys19:survey-list' %}">更新調查表</button>
                     </div>
                 </div>
